ref: move command utilities
