+++ src/CoreLib/Game/Commands/CommandBanProtocol.cs
@@ -27,7 +27,7 @@ internal class CommandBanProtocol : CommandProtocol {
         }
 
         // The time will be in the format of 1d2h3m4s. Parse it into a TimeSpan.
-        if (!TryParseDuration(time, out var timeSpan)) {
+        if (!CommandUtilities.TryParseDuration(time, out var timeSpan)) {
             InformSenderClient("Invalid time format");
             return;
         }
@@ -62,7 +62,7 @@ internal class CommandBanProtocol : CommandProtocol {
         }
 
         // The time will be in the format of 1d2h3m4s. Parse it into a TimeSpan.
-        if (!TryParseDuration(time, out var timeSpan)) {
+        if (!CommandUtilities.TryParseDuration(time, out var timeSpan)) {
             InformSenderClient("Invalid time format");
             return;
         }
@@ -84,7 +84,7 @@ internal class CommandBanProtocol : CommandProtocol {
         }
 
         // The time will be in the format of 1d2h3m4s. Parse it into a TimeSpan.
-        if (!TryParseDuration(time, out var timeSpan)) {
+        if (!CommandUtilities.TryParseDuration(time, out var timeSpan)) {
             InformSenderClient("Invalid time format");
             return;
         }
@@ -115,31 +115,5 @@ internal class CommandBanProtocol : CommandProtocol {
         sb.AppendLine($"Last infraction: {account.InfractionHistory.LastInfractionTime}");
 
         InformSenderClient(sb.ToString(), true);
-    }
-
-    private static bool TryParseDuration(string durationString, out TimeSpan result) {
-        result = TimeSpan.Zero;
-
-        // Use regular expression to match and extract components
-        var match = Regex.Match(durationString, @"(\d+d)?(\d+h)?(\d+m)?(\d+s)?");
-
-        if (match.Success) {
-            // Try to extract and convert each component
-            if (match.Groups[1].Success && int.TryParse(match.Groups[1].Value.TrimEnd('d'), out int days))
-                result += TimeSpan.FromDays(days);
-
-            if (match.Groups[2].Success && int.TryParse(match.Groups[2].Value.TrimEnd('h'), out int hours))
-                result += TimeSpan.FromHours(hours);
-
-            if (match.Groups[3].Success && int.TryParse(match.Groups[3].Value.TrimEnd('m'), out int minutes))
-                result += TimeSpan.FromMinutes(minutes);
-
-            if (match.Groups[4].Success && int.TryParse(match.Groups[4].Value.TrimEnd('s'), out int seconds))
-                result += TimeSpan.FromSeconds(seconds);
-
-            return true;
-        }
-
-        return false;
     }
 }
